fix(api): keep the top-level validateAuth guard as the type carrier
The per-route validateAuth literals broke sibling macros' type
composition (invalidateQuery's body-schema merge disappeared from the
Eden surface, failing the web typecheck). Restoring the top-level
guard reinstates the pre-branch type topology; the per-request
memoization keeps its extra resolve hook at zero additional queries
(floor re-verified at 6). The security test keeps the load-bearing
direction only: every route in those files must declare permissions.
